Update README.md
The Forest Resources website is an informative platform designed to educate and raise awareness about the importance of forests and their resources. This project aims to provide comprehensive information on various aspects of forests, including their biodiversity, conservation efforts, sustainable management practices, and the benefits they offer to the environment and society.